requirements.txt
langchain==0.2.16
langchain-openai==0.1.25
langchain-community==0.2.16
gradio==4.44.0
python-dotenv==1.0.1
pypdf==4.3.1
python-docx==1.1.2
faiss-cpu==1.8.0.post1
numpy==1.26.4
